added AddTodo file and fixed Todolist file
