+++ src/AddTodoForm.js
@@ -1,0 +1,16 @@
+import React from 'react';
+
+function AddTodoForm () {
+  return (
+    <div>
+      <form>
+        <label htmlFor="todoTitle"> Title </label>
+        <input type="text" id="todoTitle"></input>
+        <button type="submit"> Add </button>
+      </form>
+    </div>
+  )
+}
+
+
+export default AddTodoForm;
